Сократил количество шагов в тестах трансфера, депозита и кредита

+++ src/main/api/models/create_creadit_response.py
@@ -5,7 +5,7 @@ from src.main.api.models.base_model import BaseModel
 
 class CreateCreditResponse(BaseModel):
     id: int
-    amount: int
+    amount: float
     termMonths: int
-    balance: int
+    balance: float
     creditId: int
